Move classes in alvr_server.cpp into corresponding files.
